pyproject.toml
[tool.poetry]
name = "TaskView"
version = "0.1.0"
description = ""
authors = ["virusxdRJ <@joaooliveira247>"]

[tool.poetry.dependencies]
python = "^3.9"
psutil = "^5.9.1"
typer = "^0.4.1"
py-dashing = "^0.3.dev0"
black = "^22.6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
taskviewer = "taskview.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
